refactor(build): netlify serve
